feat: Add typedoc annotations to custom types & refactor export structure
