+++ src/types/tokens.ts
@@ -1,33 +1,64 @@
 import { SignableToken } from '../api';
 
+/**
+ * An ERC20 token
+ */
 export interface ERC20Token {
   type: 'ERC20';
   tokenAddress: string;
 }
 
+/**
+ * An ERC721 token
+ */
 export interface ERC721Token {
   type: 'ERC721';
   tokenId: string;
   tokenAddress: string;
 }
 
+/**
+ * An ETH token
+ */
 export interface ETHToken {
   type: 'ETH';
 }
 
+/**
+ * An amount of ETH token of unit Wei
+ */
 export interface ETHAmount extends ETHToken {
+  /**
+   * An amount in unit Wei
+   */
   amount: string;
 }
 
+/**
+ * The token details and amount of ERC20 token units
+ */
 export interface ERC20Amount extends ERC20Token {
+  /**
+   * An amount in ERC20 unit
+   */
   amount: string;
 }
 
+/**
+ * Union type that represents all token types
+ */
 export type AnyToken = ETHToken | ERC721Token | ERC20Token;
 
+/**
+ * Union type that represents all token type amounts
+ */
 export type TokenAmount = ETHAmount | ERC20Amount | ERC721Token;
 
-//SignableToken on the API requires fields in snake_case format
+/**
+ * Helper method to convert token type to a SignableToken type
+ * @param token - the token type to convert to a SignableToken type
+ * @returns the converted SignableToken
+ */
 export function convertToSignableToken(token: TokenAmount): SignableToken {
   switch (token.type) {
     case 'ERC721':
@@ -54,4 +85,3 @@ export function convertToSignableToken(token: TokenAmount): SignableToken {
       };
   }
 }
-
